Replace server name plpsipp1v with ps1images
plpsipp1v.stsci.edu is being retired soon and has been replaced by a new
server.  The name ps1images.stsci.edu (which also worked for the old
server) should be used instead.  ps1images.stsci.edu will continue to
work as long as the PS1 image cutout server is running.

These changes should be propagated to users ASAP -- the old server will
be shut down within about the next 2 weeks, at which point the panstamps
package will stop working.

It would also be desirable to use https instead of http to access the
services on ps1images.  That change would require more testing to
confirm that the access mechanisms continue to work.  This version of
the change is simpler.